requirements.txt
Flask
Flask-SQLAlchemy
Flask-WTF
Flask-Migrate
Flask-Mail
PYJWT
dnspython
gunicorn
python-whois
email_validator
psycopg2
pandas
